Add BPM utility scripts to support additional firmware update requirements
Due to the additional requirements with BPM firmware update 1.05, SMART
provided several utility scripts to meet these new requirements. Two of
those are of use to hostboot. insertBpmFwCrc.py and imageCrc.c, which
will be used to calculate and add a CRC to the flash image prior to
pre-processing the flash image into a binary.

Change-Id: I9cd751f286b039b67e29ea8f5c7b69b1afa85cb9
RTC:212446
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/80197
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>