Make RT_TARG id generation code common between IPL time and runtime
For axone we are writing the OMI mmio bars into hdat so the hypervisor
know how to talk to the devices. IPL code needs to be able to lookup
the hbrt-style ids so we can use them to make hdat entries that the
hypervisor will be able to associate targets with. This commit also
move rt_targeting.H to the correct include directory and updates
everywhere that it is included.

Change-Id: I31deaa1a9c5a7523622a8b3b12ad459e2b2feed3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/80419
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>