Remove CONFIG_AXONE_BRING_UP from xscomtest
Initially tried with local checkstop register,
but that register has simics actions which will
clear the set bit0. This was previously incorrectly
setting the wrong bit, but FIR action was executed
and setting bit0.  To make things cleaner a new
FIR register was chosen that doesn't have simics
actions associated with it.

Change-Id: Iae2d429b82b8c3eab94db2f46495a1b488347b15
RTC:201738
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/80426
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>