Handle processor swap between slots to 1-socket system
If a processor was booted in the second slot, it will be programmed
to use the memory for that slot.  When it is installed in the first
slot it will then get reprogrammed to use the data for slot0.
However, if the new system only contains data for that 1 slot, we
won't be able to find a match to do the initial part of the boot.

This change will force some values into good enough shape to get
the boot far enough to do the SBE update to reprogram the memory
map.

Change-Id: I9b88d4181272104a8c680e9b5e84c4d204fdea05
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84176
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>