Account For TPM Not Required in Node Comm
When the TPM required policy is off, node comm path should
not commit any errors as to not confuse the testers/customers.
This commit silences the errors from the node comm protocol
when TPM is not required. In addition, when TPM is required,
we want to fail the boot if node comm is unsuccessful. Logic
has been added to pass the error info up to FSP in case
of any error in istep18 (when node comm runs).

Change-Id: I76d48780790b6fee759b3ccc60a858f0a8835dee
CQ: SW466658
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81376
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>