Always position HOMER in HV space for istep 16
P9N/P9C always inits to DD2.x mode for hostboot
which doesn't support SMF.  This means that the
HCODE non secure transition to HCODE secure won't
work.

Because of that, we now force istep 16 to always place HOMER in
non SMF memory (in UV mode if SMF active) and
transition into UV there.  Then in istep 21
the HOMER is rebuilt correctly for SMF.  Note
that all the changes are in istep 15 as istep 21
uses the ATTR setup there.

Change-Id: I566903dc373667a410ddcc0a1d2cc91bcd94b08d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81202
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>