OR mux select with the "enable" bit 0b1000 when making selection
During simics bringup we figured out that the mux selects needed to
be between 8-15. We learned during bringup this is because the real
0-7 select needs to be OR'ed with the enable bit for the model of
mux. This commit does the ORing correctly and subtracts 8 from all
of the muxSelects used in the simics xml so we match what the MRW
is doing.

Change-Id: I45faab455afdfc63ac05fc2637890f9d0137a444
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81397
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>