NVDIMM: Fix issue where a deviceRead of a register is returning bad data
The reason the deviceRead is returning bad data, from an NVDIMM register,
is because the register data is segmented across several pages.  One must
first set the page correctly before attempting to do a read of a register.
This fix will set the page up correctly before doing a read/write to an
NVDIMM register.

Change-Id: Ic63c8671c0083f1ca243f255a29d1ef4606b7ce5
CQ: SW465917
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81542
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>