Port P10 HB compile flag mss error, update p10_lib mk
Top level template is NOT unique between the two files
when using the -Os compile flag. The linker just picks
the wrong file template definition.
In -O3 everything is left in place, inline so it just works
which is why we didn't hit this error before in HB

Updated makefile to exclude all the explorer UTs
to isolate changes to build and update in a
single commit.  Here we updated p9a to p10
eff_config and freq HWPs and dependent libs

Change-Id: Iab06055181ce8f67eda581b4405aadac69536e83
Original-Change-Id: Ibb359cdd6d5dd347442cf8c45ef9542bd67e7a98
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81718
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mark Pizzutillo <mark.pizzutillo@ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82464
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>