Fix BPM script to put temp files into obj tree
Depending on where the input files and scripts live and if we are
compiling in a backing build or a sandbox, the temporary (.crc)
files can sometimes be created inside a readonly tree.  This change
will place the temp files in the same (writeable) output directory
as the final binary images.

Change-Id: Ib5e6ed5ffeb825d46681eda89d93311931365663
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81787
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>