Fix BPM flash image script to properly calculate line length
Before this commit, the flash image script made certain assumptions
about the flash image file when calculating the number of bytes on a
line. Those assumptions aren't always true. Therefore, this commit adds
a function to properly determine the number of bytes on a line for any
arbitrary amount of extraneous characters on the line by stripping them
all off prior to calculation.

Change-Id: Ibeea1f62177109a512920be1434557ac55510743
RTC:212448
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81947
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>