Increased the BMC timeout from 5 seconds to 30 seconds
On a zaius machine, the BMC was consistently hitting a checkstop
with the system shutting down with error status 0x613 following
istep 13.11 - mss_draminit_training.

The theory was that it was hitting a timeout issue.  Increasing
the timeout from 5 seconds to 30 seconds abated the situation. It
does appear that it was timeout issue.

Change-Id: I825f461855b2a4a70c732f0dda308b2579d60a7d
CQ:SW466685
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82034
Reviewed-by: Glenn Miles <milesg@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>