Do not unload modules loaded for unit tests
For some of the unit tests we must load in .so modules that were
unloaded during the boot. Modules must get loaded and unloaded during
the boot, especially before we expand from cache containment to over
come size limitations. We were seeing issues were test case A and
B both relied on a module and attempted to load it when the test
case was instantiated then tried to unload the module when the test
case was completed. This was causing issues if two tests were using
the same loaded module and one test finished early and unloaded it.
Test cases are run on simics after memory is expanded so there is
no reason unload the extra modules we load in so we will leave them
loaded.

Change-Id: Ia41d38da11400f54ee2e59e497b9610ac24f1629
RTC: 214302
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/91542
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>