Use REL_POS of OMI and not OCMB for calculating hdat DIMM ID's
There is a bug in the code where we are attempting to lookup the
REL_POS of an ocmb target while trying to generate the id hdat will
use to uniquely identify a given dimm target. OCMB targets do not
have the REL_POS attribute and if they did it would always be 0,
or it would match the associated DIMM's REL_POS (for single dimm
ocmbs at least). Instead we will use the OMI's REL_POS which should
work because OMI's are 1-1 with OCMB targets.

Change-Id: Ib22307c7f4c2d97a5368fd507fe7b4965135c76e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82120
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Jayashankar Padath <jayashankar.padath@in.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>