Introduce exp_getidec HWP
This hardware procedure is used to read out the chipId and the EC
level of a given explorer ocmb. The chipId is an IBM standard that
provides information about what kind of chip it is and where it was
made. The EC is the write level of the chip. The initial plan was to
read all of this info from the CHIP_INFO register but we found that the
2nd nibble of the EC level , which represents the "minor" ec level, was
not populated correctly so we must instead read that value from the
EFUSE_OUT3 regsiter.

Change-Id: Id3dca4449c010ad42f7f3d55428b4efb3f6dfbf0
RTC: 214627
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82712
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: Mark Pizzutillo <mark.pizzutillo@ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84534
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84541
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>