Add "Not Wired" option for OPTICS_CONFIG_MODE attrs
For Swift systems obus1 is not wired to anything. This is causing
issues when we attempt to enable the link. If we set it to NOT_WIRED
then we should avoid the problem we are seeing currently where obus1
is having the NV link enabled which is getting parity errors as soon
as we enabled the obus.

Change-Id: I1380388cbf9a967380c6657102a1b1ae9b266014
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82766
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82794
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>