Remove secure boot compile option from bootloader
Make bootloader code compile one way, so that regardless of
a secureboot compiled bootloader or not, the run will succeed.
Right now, running a bootloader without secureboot compiled in
with hostboot code that is compiled with secureboot will result
in a fail. Solution is to just run through the bootloader
secureboot path regardless of having secureboot on or off.

Change-Id: I06e0ecc541db388fe96e9ae2193e0bf7e4661b9f
RTC: 211296
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82770
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>