Added code to support doing an NVM health check
Added a method that will do an NVM (non-volatile memory)
health check.  In particular this method will check the
flash error counts and does a predictive callout
if the number of flash error counts exceeds the maximum
allowed.  This method also checks the bad block percentage
and does a predictive callout if the number of bad block
percentage exceeds the maximum allowed.  A predictive
callout is done if either or both fail the check.

Added support in the runtime commands to make the nvm health check
call as well.

Also, when I did the ES (energy source) health check method, I was
not very explicit in the method that it was doing an ES health
check.  So I updated the verbiage in the nvDimmCheckHealthStatus
to add ES wherever appropriate as to make these two method
explicit as to what health check is being performed.

Change-Id: Ib9925fd2bb8430cf2121108329247d96072beb1b
CQ: 473220
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82843
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Glenn Miles <milesg@ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>