Add more FFDC to BPM Update Error logs
Adds part callouts for all BPM related error logs and collects BPM
traces for any errors that are returned from interactions with the
NVDIMM to provide context. Additionally, if there is an error during the
config updates the code will erase the firmware section to force an
update on any subsequent attempt.

Change-Id: I045b4ba1feea2fbc1ddb5e418bf60b86d6c11bb1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/82946
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>