Add wrapper to exp_getidec HWP so we can call it in platform code
There has been a trend to create more HWPs that process raw data.
This allows Hostboot and Cronus to share more code. The trick is
that we generally want to call these HWPs inside traditionally
hostboot-only code. This new fapiwrap directory will help encapsulate
all of the calls to ekb code in one place. That way if we ever need
to remove HWPs/EKB code it will be easier to stub out the calls. In the
past we have contaminated a lot of Hostboot-only code with calls out
to EKB code. This new new directory should help avoid this issue.

Change-Id: I2b2d0d1e62c97f0976c9c9ede3fe2eac30c7a40f
RTC: 214627
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83029
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84072
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>