Update SYS's ATTR_XSCOM_BASE_ADDRESS in SMF Paths
The correct value of ATTR_XSCOM_BASE_ADDRESS was not being propagated
to HDAT when SMF is enabled. The problem is that the system target has
a copy of said attribute that needs to be updated with the SMF bit
(bit 15) to reflect the fact that XSCOM BAR is in SMF memory. The
attribute under the system target was not being updated accordingly
(only those attributes under proc targets were). This commit adds a
fix to update ATTR_XSCOM_BASE_ADDRESS under the system target in SMF
mode.

Change-Id: Icac5bb7023ac0ecc521a961681420bfc21a46eac
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83057
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>