Add call to fapi wrapper of exp_getidec into hwasPlat code
The file hwasPlat.C contains the code we use to lookup and compare
the IDEC information of a given OCMB when read from the card's SPD
versus what is read from the card's defined registers. Recently the
code that looks up the IDEC from the card's registers was made into
a HWP so it could be leveraged by multiple platforms. A new wrapper
has been created on the hostboot side in src/usr/fapiwrap/ that provides
an interface for the hostboot platform to call in order to leverage
the new EKB hwp.

Change-Id: I9155767b199c9b197c010521f33cc0d418901628
RTC: 214627
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83109
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>