Update DIMM mux_select to match their OCMB counterparts
During swift bringup we realized that we had the MUX_SELECT wrong on
all of the OCMB/DIMM targets in our simics xml. Basically we were
OR'ing a 0x08 to the value that we should not have. The fix was to
add the OR'ing of the 0x08 to the code, and subtract 0x08 from what
we had set in the xml we use for simics. The net result was nothing
but this makes our XML more correct. For some reason this change
was not propgated to the dimms. This commit subtracts 0x08 from
the MUX_SELECT of the dimms in the simics xml for Axone.

Change-Id: I915d1f51d66bf0089346afa7f8c491de69bb1498
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83205
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>