Changed the callout in the call NVM Health Status to be an NVDIMM
In the call nvDimmNvmCheckHealthStatus, the call is incorrectly using
HB when it should be associated with NVDIMM.  Switched out the
addPartCallout with HWAS::BPM_PART_TYPE to a addHwCallout with the
given NVDIMM.

Also added more trace info to print out the registers that are being
retrieved.

Corrected the logic such that a good state passes.  Apparently, the
way the code was written, a good or bad state, both would fail.  Now
the good state passes and the bad state fails.

Change-Id: I46b62ce3fe5c5ac86efb501e2a91fdea8160b9b5
CQ: SW475265
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83673
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>