Retrieve LCOV from Github when building a code coverage report
This commit adds an LCOV submodule to Hostboot for the LCOV tool.
Whenever users run "make lcov", the lcov submodule will automatically
be cloned if necessary before attempting to generate a code coverage
report. This ensures that the appropriate version of lcov is available
even on older machines.

The LCOV submodule must be explicitly initialized; it will not be
cloned when cloning the Hostboot repository, even if cloning
submodules recursively.

Tested with GCC 4.9.2 and 8.2.0.

Change-Id: Ia884e0ed0e356aa9f7c99e8191ce7ab325d82dac
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83669
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Luis P Fernandez <luis.fernandez@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>