Support ATTR_PG and NMMU power-gating rules
This commit adds support for the PG attribute, which is new in P10.

ATTR_PG is a 32-bit integer present on each PERV target. The structure
of its value is the same as that of the PG MVPD row for the chiplet
related to the given PERV. The bits in the value represent the
functional state of the corresponding part. Unlike ATTR_PG_MVPD, whose
value indicates whether the part is functional according only to the
partial-good data in MVPD, the value of each bit in ATTR_PG reflects
the combined result of all methods of deconfiguration (GARD records,
deconfiguration rules, PG MVPD, etc) and indicates whether the part
has been marked non-functional for any reason by any component.

ATTR_PG is used by hardware procedures to test parts for functionality.

This commit also adds support for the fact that NMMU 1 will be
power-gated if PAUs 0, 4 and 5 are deconfigured.

Also fixes a bug in snprintf where the output string would be
unterminated if the buffer is too small.

Change-Id: Ia01dc6c44b589044bb88aaf3d21e4e820d92638d
RTC: 213276
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83716
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Glenn Miles <milesg@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>