Fix trace string hashing collision in HB
When creating hash values for trace strings in p10_mss_eff_grouping.C
collisions were found inside displayMemoryBarRegs() for tracing
MCFGP1/MCFGP0A and MCFGPM1/MCFGPM0A values.  This commit makes slight
changes to the strings in order to prevent the hash collisions as
a quick-fix.  In addition to this quick-fix, a story will be opened
to investigate using a better hashing algorithm.

Change-Id: I90e7d1da69922cda84ee574de5c996a4db2896d2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83722
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83727
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>