Hostboot platform support for Explorer inband commands via i2c
Inband SRAM can be accessed via scom i2c commands.  To Explorer,
a register address and an internal memory address are the same thing.
That allows us to execute the inband command set even if the
OMI link is not active.  A new attribute determinds when this
inband i2c is required so it can be an easy override for lab use.
By default, this i2c path will not execute when OMI links are working.

Change-Id: I3f18cf78d2e88e33935f1bd241ef4e2796d36d93
RTC: 208447
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83787
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>