Update axone pnor layout to move eecache to the begining
Because we rely on having a stable offset for the EECACHE section
of pnor we will move it to the start of the pnor layout. This will
make it less likely to move around as now we are auto-adjusting
offsets if sections like HBI happen to exceed to allotted space.

Change-Id: I4463c9d1cfb6ba86c777c976f911b14be5a483e3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83791
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>