Add new DDR4 DDIMM keyword ENTIRE_SPD_WITHOUT_EFD
With the DDR4 DDIMM SPD there are really two parts. There is the
640 bytes of information you typically see in a DIMM's SPD and then
there are 2 KB - 640 bytes of EFD (extensible function descriptors)
to give vendors space to add customized data. In many cases we want
the ENTIRE_SPD (including the EFD area) but in other cases we want
to save some time/space and only read in the first 640 bytes. This
new keyword will give us that flexibility.

Change-Id: I88ed58b37417b576f2ec4a779ee0c114791c6c21
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84070
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>