PRD: NVDIMM workaround for ES TEMP glitches
There is currently a glitch with the energy source temperature
readings. We are getting intermittent incorrect readings for
the temperature that is causing us to analyze to an ES temperature
warning or error because of these glitched readings. This is a
workaround to make the ES temp logs hidden until we hit
threshold. The fix for the root cause of the glitches will
need to come from SMART.

Change-Id: I8df3f2531fda2e4c0273bfe3714c7efb9662971d
CQ: SW476951
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84345
Reviewed-by: Paul Greenwood <paul.greenwood@ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84363
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84344
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/87521
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/87647
Tested-by: Zane C Shelley <zshelle@us.ibm.com>