Make snprintf always terminate the output buffer
Previously, the snprintf function would not terminate the output
buffer if it had already filled the entire buffer with
characters. This would lead to the creation of unterminated strings
which could result in buffer overreads and/or information
disclosures.

This commit fixes that issue, making it conform to the C99 standard
(section 7.19.6.5).

Also addresses a buffer underflow that could occur when the source
string to an sprintf call contains too many \b characters.

Change-Id: Ie516b1c7d74d37cc5f48fe03693f096fe4bd6c02
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/84325
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Luis P Fernandez <luis.fernandez@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>